+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added async support for Tokio runtime.
+
 ### Fixed
 
 - Added missing method comments to improve documentation clarity.